temp fix to plans page background not scrolling
Temporary bug fix to plans page, where the background stops at the bottom of the screen. Normally is only a problem when more than 6 plans are added, where in the screen scrolls down and the background ends. Ideally, we would want the window height to update the background as more and more plans are added, but a cap of 6 would also fix this.